trim carriage-return at the end of signature comments
This commit fixes a bug when parsing minisign signatures.
Before, any carriage return ('\r') at the end of a
trusted / untrusted comment was not removed. This lead to
signature verification failures. In particular, the verification
of the comment signature fails when e.g. the OS uses `\r\n` for
newlines instead of `\n`.

Byte representation of a trusted comment - with and without the
trailing carriage return.
```
[116 105 109 101 115 116 97 109 112 58 49 54 49 55 55 50 49 48 50 51 9 102 105 108 101 58 114 111 98 116 101 115 116 46 112 115 49 13]
[116 105 109 101 115 116 97 109 112 58 49 54 49 55 55 50 49 48 50 51 9 102 105 108 101 58 114 111 98 116 101 115 116 46 112 115 49]
```

Fixes #7